refactor: extract reuse method to helpers
