+++ spec/app_info/android/signatures/v1_spec.rb
@@ -3,11 +3,12 @@ describe AppInfo::Android::Signature::V1 do
   let(:signature_description) { AppInfo::Android::Signature::V1::DESCRIPTION }
 
   context 'when parse v1 signature apk' do
-    let(:file) { fixture_path('apps/android.apk') }
+    let(:file) { fixture_path('apps/wear.apk') }
     let(:parser) { AppInfo.parse(file) }
-    subject { AppInfo::Android::Signature::V1.verify(1, parser) }
+    subject { AppInfo::Android::Signature::V1.verify(parser) }
 
-    it { expect(subject.scheme).to eq('v1') }
+    it { expect(subject.version).to eq(version) }
+    it { expect(subject.scheme).to eq("v#{version}") }
     it { expect(subject.description).to eq(signature_description) }
     it { expect(subject.signatures).to be_kind_of(Array) }
     it { expect(subject.signatures).not_to be_empty }
@@ -20,9 +21,9 @@ describe AppInfo::Android::Signature::V1 do
   context 'when parse v2 signature only apk' do
     let(:file) { fixture_path('apps/android-v2-signed-only.apk') }
     let(:parser) { AppInfo.parse(file) }
-    subject { AppInfo::Android::Signature::V1.verify(version, parser) }
+    subject { AppInfo::Android::Signature::V1.verify(parser) }
 
-    it { expect { AppInfo::Android::Signature::V1.verify(version, parser) }.to raise_error(AppInfo::Android::Signature::SecurityError) }
-    it { expect { subject.verify }.to raise_error(AppInfo::Android::Signature::SecurityError) }
+    it { expect { AppInfo::Android::Signature::V1.verify(parser) }.to raise_error(AppInfo::Android::Signature::NotFoundError) }
+    it { expect { subject.verify }.to raise_error(AppInfo::Android::Signature::NotFoundError) }
   end
 end
