+++ spec/app_info/android/signatures/v2_spec.rb
@@ -5,17 +5,18 @@ describe AppInfo::Android::Signature::V2 do
   context 'when parse v1 signature apk' do
     let(:file) { fixture_path('apps/wear.apk') }
     let(:parser) { AppInfo.parse(file) }
-    subject { AppInfo::Android::Signature::V2.new(version, parser) }
+    subject { AppInfo::Android::Signature::V2.new(parser) }
 
-    it { expect { AppInfo::Android::Signature::V2.verify(version, parser) }.to raise_error(AppInfo::Android::Signature::SecurityError) }
-    it { expect { subject.verify }.to raise_error(AppInfo::Android::Signature::SecurityError) }
+    it { expect { AppInfo::Android::Signature::V2.verify(parser) }.to raise_error(AppInfo::Android::Signature::NotFoundError) }
+    it { expect { subject.verify }.to raise_error(AppInfo::Android::Signature::NotFoundError) }
   end
 
   context 'when parse v2 signature apk' do
     let(:file) { fixture_path('apps/android-v2-signed-only.apk') }
     let(:parser) { AppInfo.parse(file) }
-    subject { AppInfo::Android::Signature::V2.verify(version, parser) }
+    subject { AppInfo::Android::Signature::V2.verify(parser) }
 
+    it { expect(subject.version).to eq(version) }
     it { expect(subject.scheme).to eq("v#{version}") }
     it { expect(subject.description).to eq(signature_description) }
     it { expect(subject.certificates).to be_kind_of(Array) }
@@ -31,9 +32,9 @@ describe AppInfo::Android::Signature::V2 do
   context 'when parse v3 signature apk' do
     let(:file) { fixture_path('apps/android-v3-signed-only.apk') }
     let(:parser) { AppInfo.parse(file) }
-    subject { AppInfo::Android::Signature::V2.verify(version, parser) }
+    subject { AppInfo::Android::Signature::V2.verify(parser) }
 
-    it { expect { AppInfo::Android::Signature::V2.verify(version, parser) }.to raise_error(AppInfo::Android::Signature::SecurityError) }
-    it { expect { subject.verify }.to raise_error(AppInfo::Android::Signature::SecurityError) }
+    it { expect { AppInfo::Android::Signature::V2.verify(parser) }.to raise_error(AppInfo::Android::Signature::NotFoundError) }
+    it { expect { subject.verify }.to raise_error(AppInfo::Android::Signature::NotFoundError) }
   end
 end
